+++ dotcom-rendering/src/types/content.ts
@@ -1,10 +1,6 @@
 // -------------------------------------
 // Elements
 
-import type { PersonalityQuizAtom } from '../components/PersonalityQuizAtom.importable';
-
-export type QuizAtomType = Parameters<typeof PersonalityQuizAtom>[0];
-
 // -------------------------------------
 interface ThirdPartyEmbeddedContent {
 	isThirdPartyTracking: boolean;
@@ -134,7 +130,7 @@ interface QuizAtomBlockElement {
 	quizType: 'personality' | 'knowledge';
 	id: string;
 	questions: QuestionType[];
-	resultBuckets: QuizAtomType['resultBuckets'];
+	resultBuckets: ResultsBucketType[];
 	resultGroups: {
 		id: string;
 		title: string;
@@ -866,7 +862,7 @@ export type MessageUsFieldType = CampaignFieldText | CampaignFieldTextArea;
 // Quiz
 // -------------------------------------
 
-type AnswerType = {
+export type AnswerType = {
 	id: string;
 	text: string;
 	revealText?: string;
@@ -874,12 +870,59 @@ type AnswerType = {
 	answerBuckets: string[];
 };
 
-type QuestionType = {
+export type QuestionType = {
 	id: string;
 	text: string;
 	answers: AnswerType[];
 	imageUrl?: string;
 	imageAlt?: string;
+};
+
+export type ResultGroupsType = {
+	title: string;
+	shareText: string;
+	minScore: number;
+	id: string;
+};
+
+export type ResultsBucketType = {
+	id: string;
+	title: string;
+	description: string;
+};
+
+export type KnowledgeQuizAtomType = {
+	id: string;
+	questions: QuestionType[];
+	resultGroups: ResultGroupsType[];
+	sharingUrls: SharingUrlsType;
+	theme: ArticleTheme;
+};
+
+export type PersonalityQuizAtomType = {
+	id: string;
+	questions: QuestionType[];
+	resultBuckets: ResultsBucketType[];
+	sharingUrls: SharingUrlsType;
+	theme: ArticleTheme;
+};
+
+export type QuizSelectionType = Record<string, AnswerType>;
+
+export type SharePlatformType =
+	| 'facebook'
+	| 'twitter'
+	| 'email'
+	| 'linkedIn'
+	| 'pinterest'
+	| 'whatsApp'
+	| 'messenger';
+
+export type SharingUrlsType = {
+	[K in SharePlatformType]?: {
+		url: string;
+		userMessage: string;
+	};
 };
 
 // -------------------------------------
